Refactor the clustered event bus tests to reduce the number of tests
