+++ src/test/java/io/vertx/test/core/ClusteredEventBusWithSSLTest.java
@@ -19,7 +19,6 @@ package io.vertx.test.core;
 import io.vertx.core.VertxOptions;
 import io.vertx.core.eventbus.EventBusOptions;
 import io.vertx.core.http.ClientAuth;
-import io.vertx.core.net.*;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
@@ -28,10 +27,11 @@ import java.util.Arrays;
 
 /**
  * Tests the clustered event bus with various SSL / TLS configuration.
+ *
  * @author Clement Escoffier
  */
 @RunWith(value = Parameterized.class)
-public class ClusteredEventBusWithSSLTest extends ClusteredEventBusTest {
+public class ClusteredEventBusWithSSLTest extends ClusteredEventBusTestBase {
 
   private final EventBusOptions options;
 
@@ -71,29 +71,22 @@ public class ClusteredEventBusWithSSLTest extends ClusteredEventBusTest {
     // Parameters:
     //KeyCert, Trust, requireClientAuth, clientTrustAll, useCrl, enabledCipherSuites
 
-    return Arrays.asList(new Object[][] {
-        { KeyCert.JKS, Trust.NONE, false, true, false, new String[0] }, // trusts all server certs
-        { KeyCert.JKS, Trust.JKS, false, false, false, new String[0] },
-        { KeyCert.JKS, Trust.JKS, false, false, true, new String[0] },
-        { KeyCert.PKCS12, Trust.JKS, false, false, false, new String[0] },
-        { KeyCert.PEM, Trust.JKS, false, false, false, new String[0] },
-        { KeyCert.JKS_CA, Trust.JKS_CA, false, false, false, new String[0] },
-        { KeyCert.JKS_CA, Trust.PKCS12_CA, false, false, false, new String[0] },
-        { KeyCert.JKS_CA, Trust.PEM_CA, false, false, false, new String[0] },
-        { KeyCert.PKCS12_CA, Trust.JKS_CA, false, false, false, new String[0] },
-        { KeyCert.PKCS12_CA, Trust.PKCS12_CA, false, false, false, new String[0] },
-        { KeyCert.PEM_CA, Trust.JKS_CA, false, false, false, new String[0] },
-        { KeyCert.PEM_CA, Trust.PKCS12_CA, false, false, false, new String[0] },
-        { KeyCert.PEM_CA, Trust.PEM_CA, false, false, false, new String[0] },
-        { KeyCert.JKS, Trust.PKCS12, false, false, false, new String[0] },
-        { KeyCert.JKS, Trust.PEM, false, false, false, new String[0] },
-        { KeyCert.JKS, Trust.JKS, true, false, false, new String[0] },
-        { KeyCert.JKS, Trust.PKCS12, true, false, false, new String[0] },
-        { KeyCert.JKS, Trust.PEM, true, false, false, new String[0] },
-        { KeyCert.PKCS12, Trust.JKS, true, false, false, new String[0] },
-        { KeyCert.PEM, Trust.JKS, true, false, false, new String[0] },
-        { KeyCert.PEM_CA, Trust.PEM_CA, true, false, false, new String[0] },
-        { KeyCert.JKS, Trust.PEM_CA, false, true, false, HttpTest.ENABLED_CIPHER_SUITES },
+    return Arrays.asList(new Object[][]{
+        {KeyCert.JKS, Trust.NONE, false, true, false, new String[0]}, // trusts all server certs
+        {KeyCert.JKS, Trust.JKS, false, false, false, new String[0]},
+        {KeyCert.JKS, Trust.JKS, false, false, true, new String[0]},
+        {KeyCert.PKCS12, Trust.JKS, false, false, false, new String[0]},
+        {KeyCert.PEM, Trust.JKS, false, false, false, new String[0]},
+        {KeyCert.JKS_CA, Trust.JKS_CA, false, false, false, new String[0]},
+        {KeyCert.PKCS12_CA, Trust.JKS_CA, false, false, false, new String[0]},
+        {KeyCert.PEM_CA, Trust.PKCS12_CA, false, false, false, new String[0]},
+        {KeyCert.JKS, Trust.PKCS12, false, false, false, new String[0]},
+        {KeyCert.JKS, Trust.PEM, false, false, false, new String[0]},
+        {KeyCert.JKS, Trust.PKCS12, true, false, false, new String[0]},
+        {KeyCert.JKS, Trust.PEM, true, false, false, new String[0]},
+        {KeyCert.PKCS12, Trust.JKS, true, false, false, new String[0]},
+        {KeyCert.PEM_CA, Trust.PEM_CA, true, false, false, new String[0]},
+        {KeyCert.JKS, Trust.PEM_CA, false, true, false, HttpTest.ENABLED_CIPHER_SUITES},
     });
   }
